Move writer process into herder daemon module
