+++ src/herder_daemon/ipc.rs
@@ -1,6 +1,8 @@
 use serde::{Deserialize, Serialize};
 
-use crate::writer_process::ipc::WriterProcessConfig;
+pub use super::writer_process::ipc::{
+    ErrorType, InitialInfo, StatusMessage, WriterProcessConfig,
+};
 
 /// Tell the herder to start a herd for performing an arbitrary action.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
